+++ tests/xua_hw_tests/test_dfu/src/vendorrequests.c
@@ -1,3 +1,5 @@
+// Copyright 2025 XMOS LIMITED.
+// This Software is subject to the terms of the XMOS Public Licence: Version 1.
 #include <stdio.h>
 #include <stdint.h>
 #include <stdlib.h>